♻️ refactor: remove context api useUser hook, add zustand useUserStore hook
