+++ src/hooks/useSpaceQuery.tsx
@@ -1,13 +1,13 @@
 import { useQuery } from "@tanstack/react-query";
 
-import useUser from "./useUser";
+import { useUser } from "./useUserStore";
 import APIClient from "utils/api";
 
 import type { Space } from "types/api";
 import type { User } from "types/auth";
 
 export function useSpaceQuery() {
-  const { user } = useUser();
+  const user = useUser();
   const api = new APIClient()
     .setAccessToken(user?.accessToken)
     .setGithubAccessToken(user?.githubAccessToken)
